+++ src/test/java/com/clickhouse/kafka/connect/sink/junit/extension/SinceClickHouseVersion.java
@@ -1,0 +1,12 @@
+package com.clickhouse.kafka.connect.sink.junit.extension;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target({ElementType.METHOD})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface SinceClickHouseVersion {
+    String value();
+}